✨feat: 중복 닉네임 처리 구현
